Renamed files
Renamed idlcs plugin.* files to generator.* to keep it compatible
with all other language bindings.
The contents of these files was previously split between other files
and the generator file was removed, by doing this other bindings
(C++/python/???) should not require any modifications.
