+++ src/tools/idlc/tests/descriptor.c
@@ -16,7 +16,7 @@
 #include "dds/ddsc/dds_opcodes.h"
 #include "idl/string.h"
 #include "libidlc/libidlc__descriptor.h"
-#include "idlc/plugin.h"
+#include "idlc/generator.h"
 #include "test_common.h"
 
 #include "CUnit/Theory.h"
