+++ src/tools/idlc/tests/type_meta.c
@@ -21,7 +21,7 @@
 #include "dds/cdr/dds_cdrstream.h"
 #include "idl/string.h"
 #include "libidlc/libidlc__descriptor.h"
-#include "idlc/plugin.h"
+#include "idlc/generator.h"
 #include "test_common.h"
 
 #include "CUnit/Theory.h"
